+++ nix/pkgs/omnixy-scripts/default.nix
@@ -4,7 +4,7 @@
 }:
 stdenvNoCC.mkDerivation {
   pname = "omnixy-scripts";
-  version = "0.1.0";
+  version = "0.1.1-preview";
 
   src = ../../..; # repo root